chore: remove unused counter and shop related JavaScript files to clean up the codebase and improve maintainability refactor: update import paths for createElement function to use the main.js file instead of the removed upgrade.js file
