--- assets/js/counter/counter.js
@@ -1,13 +1,0 @@
-let counter = 0;
-
-export function incrementScore() {
-    counter++;
-}
-
-export function getScore() {
-    return counter;
-}
-
-export function updateScoreDisplay(scoreElement) {
-    scoreElement.textContent = getScore().toString();
-}
