--- assets/js/shop/upgrade/upgrade.js
@@ -1,17 +1,0 @@
-import {createUpgradeCard} from "./card/card";
-
-export function createElement(tagName, classNames = []) {
-    const element = document.createElement(tagName);
-
-    element.classList.add(...classNames);
-
-    return element;
-}
-
-export function showUpgrades(upgrades) {
-    const shop = document.getElementById("shop");
-
-    for (const upgrade of upgrades) {
-        shop.appendChild(createUpgradeCard(upgrade));
-    }
-}
