+++ js/counter/counter.js
@@ -14,4 +14,3 @@ export function getScore() {
 export function updateScoreDisplay(score) {
   score.textContent = getScore().toString();
 }
-
