+++ js/main.js
@@ -2,6 +2,14 @@ import { loadUpgrades } from "./shop/api.js";
 import { showUpgrades, updateCardsState } from "./shop/upgrade/upgrade.js";
 import { incrementScore, updateScoreDisplay } from "./counter/counter.js";
 
+export function createElement(tagName, classNames = []) {
+  const element = document.createElement(tagName);
+
+  element.classList.add(...classNames);
+
+  return element;
+}
+
 const cookie = document.getElementById("cookieButton");
 const score = document.getElementById("score");
