+++ js/shop/api.js
@@ -3,6 +3,5 @@ export async function loadUpgrades() {
   if (!response.ok) {
     throw new Error("Failed to load upgrade");
   }
-
   return response.json();
 }
