+++ js/shop/upgrade/icon/icon.js
@@ -1,4 +1,4 @@
-import { createElement } from "../upgrade.js";
+import { createElement } from "../../../main.js";
 
 export function createUpgradeIcon(iconClass) {
   const iconContainer = createElement("div", ["upgradeIcon", "flex"]);
