+++ js/shop/upgrade/upgrade.js
@@ -1,12 +1,5 @@
 import { createUpgradeCard, unlockCard } from "./card/card.js";
 
-export function createElement(tagName, classNames = []) {
-  const element = document.createElement(tagName);
-
-  element.classList.add(...classNames);
-
-  return element;
-}
 const createdUpgradeCards = [];
 
 export function showUpgrades(upgrades) {
